% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_support.R
\name{as.party.ETree}
\alias{as.party.ETree}
\title{Transformation of a Exact Tree object to party object}
\usage{
\method{as.party}{ETree}(obj, nodeID = 1L, ...)
}
\arguments{
\item{obj}{tree of class \code{ETree}.}

\item{nodeID}{Node identification.}

\item{\dots}{additional arguments to be passed.}
}
\value{
object transformed to a \code{constparty} object.
}
\description{
Transformation of a Exact Tree object to party object
}
\keyword{as.party}
