% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ETree.control.R
\name{ETree.control}
\alias{ETree.control}
\title{Control Parameters for ETree Algorithm}
\usage{
ETree.control(
  measure = NULL,
  maxsize = 4,
  maxdepth = 2,
  minbucket = 10,
  minheterogeneity = 0.05,
  ncv = 10,
  cvVector = NULL,
  heterogeneityonly = NULL,
  BoundH = NULL,
  branchandbound = NULL,
  alltreesizes = FALSE,
  sortedpsplits = NULL,
  lookaheadheuristic = NULL
)
}
\arguments{
\item{measure}{0 = minimize residual Sum of Squares (continuous Y)
1 = minimize Misclassification Rate        (discrete Y)
3 = minimize the Risk Sum P(A)*R(A)        (discrete Y)
    Measure = scalar or Measure=c(3,c(Prior,LossM))}

\item{maxsize}{if null: only maxdepth restrictions are applied
if defined: the maximum number of terminal nodes for the best tree}

\item{maxdepth}{The maximum number of layers in a tree, NOT INCLUDING THE LAYER 0.
maxdepth=0 means a tree with just one node (the root node),
maxdepth=1 means one split and two terminal nodes, etc.}

\item{minbucket}{defines the minimum number of observations in a terminal node
if the number of observations in a node is equal to this minimum
the node becomes a possible terminal node; if the number of observations in a node
is smaller, the node is set illegal and disregarded.}

\item{minheterogeneity}{defines the situation when a node A is considered Very Homogeneous
and is not allowed to be split further, i.e. the node A becomes a possible terminal node
(|A| = number of objects in node)}

\item{ncv}{specify N for N-fold cross-validation
if NCV = 0, no N-fold cross-validation
if NCV<0, abs(N)-fold cross-validation ONLY (no tree estimation)
if NULL (default), NCV=0}

\item{cvVector}{<vector> process N-fold cross-validation according to the classes defined by <vector>
vector is assumed to assign each observational unit to one of
N-classes, number 1,...,N}

\item{heterogeneityonly}{(TRUE,FALSE)
Algorithm will only give Heterogeneity/Impurity not a tree table
=> setting to false will give treetable as well}

\item{BoundH}{optional bound for best tree value, for BranchandBound}

\item{branchandbound}{(TRUE,FALSE)
Algorithm will use branch and bound rules to speed up
=> for maxsize tree only, h only
=> make sure you specify boundH}

\item{alltreesizes}{(FALSE,TRUE)
algorithm will compute multiple trees of size<=size restriction}

\item{sortedpsplits}{(TRUE,FALSE)
Algorithm will try to speed up use a sorting logic.
=> suitable for continuous predictors with many categories}

\item{lookaheadheuristic}{Specify treedepth=the number of levels the heuristic will lookahead.
When growing a tree with lookahead search (heurist approach),
this specifies the maximum depth with respect to which
any local split is optimized. When setting this depth = 1, this
ammounts to convential tree growing.}
}
\value{
A list containing the options for the function ETree.
}
\description{
Various parameters that control aspects of the \dQuote{ETree} algorithm.
}
\examples{
ETree.control(measure=0, maxsize = 6, maxdepth = 4, minbucket = 5, ncv=10, alltreesizes = TRUE)

}
\seealso{
\code{\link{ETree}}
}
