% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.ETree.R
\name{plot.ETree}
\alias{plot.ETree}
\title{Visualisation of a Exact Tree}
\usage{
\method{plot}{ETree}(x, TerminalNodes = NULL, digits = 2, ...)
}
\arguments{
\item{x}{transformed tree of class \code{ExactTree}.}

\item{TerminalNodes}{Number of terminal nodes of the tree to plot.}

\item{digits}{specified number of decimal places of the splitpoints in the graph
(default is 2).}

\item{\dots}{additional arguments to be passed.}
}
\value{
A plot of the tree with the specified number of terminal nodes.
}
\description{
Visualisation of a Exact Tree
}
\examples{
\donttest{
  data(iris)
  # Fit an Exact Tree model
  controlEtree <- ETree.control(measure=0, maxsize = 4, maxdepth = 3,
  minbucket = 5, ncv=5, alltreesizes = FALSE)
  tree<-ETree(Species ~ Sepal.Length + Sepal.Width + Petal.Length + Petal.Width,
  control= controlEtree, data = iris)
  plot(tree, TerminalNodes=4)
}

}
\references{
Torsten Hothorn and Achim Zeileis (2013). partykit: A Toolkit for
  Recursive Partytioning. R package version 0.1-5.
}
\seealso{
\code{\link{ETree}},\code{\link{summary.ETree}}
}
\keyword{as.party}
\keyword{plot}
