% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prune.ETree.R
\name{prune.ETree}
\alias{prune.ETree}
\title{Pruning of a Exact Tree}
\usage{
\method{prune}{ETree}(tree, pp = 1, ...)
}
\arguments{
\item{tree}{fitted tree of the class \code{ETree}.}

\item{pp}{pruning parameter, the constant (\eqn{c}) to be used in the \eqn{c*}standard
error rule. The default value is 1.}

\item{\dots}{optional additional arguments.}
}
\value{
Returns an object of class \code{ETree}. The number of leaves of this object is
  equal to the optimally pruned size of the tree.
}
\description{
Determines the optimally pruned size of the tree by applying the one
standard error rule to the results from the bias-corrected bootstrap procedure.
}
\details{
The one standard error rule for \code{ETrees} uses the estimates of the bias-corrected
  criterion value (\eqn{C}) and its standard error for each value of \eqn{L}
  (= maximum number of leaves). The optimally pruned tree corresponds to the
  smallest tree with a bias-corrected \eqn{C} higher or equal to the maximum
  bias-corrected \eqn{C} minus its standard error.
}
\keyword{tree}
