% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.ETree.R
\name{summary.ETree}
\alias{summary.ETree}
\title{Summarizing Exact Trees Information}
\usage{
\method{summary}{ETree}(object, TerminalNodes = NULL, digits = 3, ...)
}
\arguments{
\item{object}{a \code{ETree} object. This can be the output of \code{\link{ETree}}.}

\item{TerminalNodes}{Number of terminal nodes of the tree to summarize.}

\item{digits}{specified number of decimal places (default is 3).}

\item{\dots}{optional additional arguments.}
}
\value{
prints a summarized version of the \code{ETree} output.
}
\description{
Summary method for an object of class \code{ETree}.
}
\details{
This function is a method for the generic function summary for class
  \code{ETree}. It extracts the following essential components from a \code{ETree}
  object: 1) Split information;
  2) Leaf information, and 3) CV information.
}
\examples{
\donttest{
  data(iris)
  # Fit an Exact Tree model
  controlEtree <- ETree.control(measure=0, maxsize = 4, maxdepth = 3,
  minbucket = 5, ncv=5, alltreesizes = FALSE)
  tree<-ETree(Species ~ Sepal.Length + Sepal.Width + Petal.Length + Petal.Width,
  control= controlEtree, data = iris)
  summary(tree, TerminalNodes=4)
}

}
\keyword{summary}
